requirements.txt
streamlit
requests
imdbpy